Create README.md
Here on steemit, I believe you don't have to be a whale or a dolphin to give back to the community. 

![Meme Creator_1520519846401.jpg](https://steemitimages.com/DQmUFhYsie3knKqPkvsCjzrFBzoqcbQsDMKCiFbRbhk5Avi/Meme%20Creator_1520519846401.jpg)

Today, I  am 59 days old on steemit and am so happy to be a part of the platform, thanks to my 400+ followers.

I remember how it was for me during my first few days on steemit; posting without being noticed. Most times, quality post of newbies goes unread, unrewarded and unnoticed. Hence my reason for deciding to run this promo.

**PROMO TAG**

![CTD822018215613.jpg](https://steemitimages.com/DQmNVB9vCe7J7nZuuSUqfyfrKYo8TyeQ4o4p8c1uY8WCcBr/CTD822018215613.jpg)

This promo is tagged **"the First-post-promo"** it does not have to be your "introduce yourself post" but your very first post.

**AIM**

![be822458187251.5a1c828c6db53-1.gif](https://steemitimages.com/DQmVupT3Ap9AikhPiR56gEKNhFKDSsuRPq6rt691m1QwGpn/be822458187251.5a1c828c6db53-1.gif)

The aim of this promo is to encourage newbies, to give them assurance that they are welcome here.

Getting their post resteemed and upvoted will go a long way in the journey of this newbies and make them know that steemit is really a life-changing platform!

**VISION**

![974933.gif](https://steemitimages.com/DQmP16TZxHATjNZXzqxKuTkvofKGo988ZcfeEL4WzDCSDAV/974933.gif)

The vision of this promo; is to be accepted by the steemit community.
I wish to have support and encouragement to increase the price tag so that this newbies will have not only have a token, but a great startup  point here on steemit.

**PROMO DETAILS**

![more-details-group-uno-4.gif](https://steemitimages.com/DQmaYZQhFVAbayYZ22jZ9YXe3SxmMY1j8inCmG7Kxd5vmb3/more-details-group-uno-4.gif)

* **First-post-promo** will run every Friday and Wednesday:

* Drop the link to your very **"first-post"** in the comment box.
* Reputation should be  (25 to 35).
* Post should not be more than 48hrs old.
* Make **#first-post-promo#** one of your tags not necessarily the first tag. *So Edit*
* The post should be in English.

**PRICE TAG**

![price_zpswsai1say.gif](https://steemitimages.com/DQmZB8g94qBKvAqKmjGLnS55dxtphBE7VfbE5ZcgHxYqRji/price_zpswsai1say.gif)

* The best **"First-post"** will be randomly selected after 48hrs. 
* This Newbie will receive 0.200 SBD from me.
* A resteem and an upvote of the **"First-post"**
* A follow from me. 
* And also an upvote, resteem and a comment on your recent post.

**I will upvote all links to "First-post" in the comment section, not necessarily "an introduce yourself post"**

**N: B** 
**Price money is little for now, because am a plankton and I believe I don't have to wait to be a whale or a dolphin to give back.**

If the posted payout is more, the price tag will be increased.

*With the progress of the promo and awareness by the steemit community, the price tag will increase.*

**ACKNOWLEDGEMENTS**

![thank-you-030.gif](https://steemitimages.com/DQmNdNyPBVoA2FxiaV2K5C7SUCfdZjAfY6S7pbfS9qgg5w9/thank-you-030.gif)

I am able to do this because of the 100SP delegation I got from @hitmeasap (Benefactor of the asapers) who decided to pay it forward (after getting a delegation from @fulltimegeek) to 5 steemians, of which I was privileged to be a part of.  

This delegation encouraged me and have so far increased my progress, hence I decided that I also want to start paying forward.

Am also a member of the @asapers which is a new curation team, with @insideoutlet and @shai-halud doing a great job as the moderators.


**SUPPORT**
![giphy-1-1.gif](https://steemitimages.com/DQmSVTLveBkJs5krHi4J4QXJ7ECYVwxA7Sj774XAosH4cwM/giphy-1-1.gif)

Encouragement and support will be appreciated for improvement and awareness of this promo.
If you want to encourage this course, kindly **Upvote!!!** **Resteem!!!** and **Comment!!!** to give better ideas.

*First-post-promo "Giving hope to Newbies " (A way of paying forward by Frankabelle).*

*Image credit*
[Image1](encrypted-tbn2.gstatic.com)
[Image 3](mri-s3-cdn-cf.behance.net)
[Image 4](www.fplschools.com)
[Image 5](www.groupuno.com)
[Image 6](picdesi.com)
[Image 7](blogacademyoflearning.com)